requirements.txt
langchain_community==0.3.23
langchain_core==0.3.58
langchain_openai==0.3.16
langgraph==0.4.1
pydantic==2.11.4
python-dotenv==1.1.0
streamlit==1.45.0
